// src/app/models/call.ts
export class Call {
    $key: string;
    callNo: string;
    type: string;
    categoryKey: string;
    categoryName: string;
    subCategoryKey: string;
    subCategoryName: string;
    purchasedFrom: string;
    dop: string;
    brandKey: string;
    brandName: string;
    quantity: string;
    complaint: string;
    warranty: string;
    customerKey: string;
    customerName: string;
    callStatus: string;
    technicianKey: string;
    technicianName: string;
    tat: string;

    constructor() {}
}