// src/app/models/customer.ts
export class Customer {
    $key: string;
    name: string;
    address: string;
    area: string;
    town: string;
    district: string;
    state = 'Tamilnadu';
    pincode: string;
    landmark: string;
    phoneNumber: string;
    altPhoneNumber: string;
    mailId: string;
    status: string;
    createdDate: string;
    createdBy: string;
    modifiedDate: string;
    modifiedBy: string;

    constructor() {}
}